refactor: 最終手情報を導出する関数を追加し、時間設定のユーティリティ関数を実装
